WIP - more billing tests
