+++ datarepo-clienttests/src/main/java/scripts/testscripts/baseclasses/BillingProfileUsers.java
@@ -1,9 +1,18 @@
 package scripts.testscripts.baseclasses;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+
+import bio.terra.datarepo.model.BillingProfileModel;
+import bio.terra.datarepo.model.DatasetSummaryModel;
+import bio.terra.datarepo.model.EnumerateBillingProfileModel;
 import java.util.List;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import runner.config.TestUserSpecification;
+import scripts.utils.tdrwrapper.DataRepoWrap;
+import scripts.utils.tdrwrapper.exception.DataRepoNotFoundClientException;
+import scripts.utils.tdrwrapper.exception.DataRepoUnauthorizedClientException;
 
 public class BillingProfileUsers extends runner.TestScript {
   private static final Logger logger = LoggerFactory.getLogger(BillingProfileUsers.class);
@@ -23,5 +32,115 @@ public class BillingProfileUsers extends runner.TestScript {
     ownerUser1 = testUsers.get(0);
     ownerUser2 = testUsers.get(1);
     userUser = testUsers.get(2);
+  }
+
+  protected enum RoleState {
+    NONE,
+    USER,
+    OWNER;
+  }
+
+  // Run is sequence of operations collecting their results. At the end compare the results
+  // to what is expected for their role state.
+  protected void testOperations(DataRepoWrap wrap, RoleState role, String profileId)
+      throws Exception {
+
+    // Test for ability to enumerate the profile
+    boolean foundEnumeration = false;
+    EnumerateBillingProfileModel enumeration = wrap.enumerateProfiles(0, 1000);
+    for (BillingProfileModel profile : enumeration.getItems()) {
+      if (StringUtils.equals(profileId, profile.getId())) {
+        foundEnumeration = true;
+      }
+    }
+    logger.info("Role {} enumeration {}", role, foundEnumeration);
+
+    // Test for ability to retrieve the profile
+    boolean retrieveSuccess;
+    try {
+      wrap.retrieveProfile(profileId);
+      retrieveSuccess = true;
+    } catch (DataRepoUnauthorizedClientException ex) {
+      retrieveSuccess = false;
+    }
+    logger.info("Role {} retrieve {}", role, retrieveSuccess);
+
+    // Test for ability to use the profile to create a dataset
+    boolean datasetSuccess;
+    try {
+      DatasetSummaryModel datasetSummaryModel =
+          wrap.createDataset(profileId, "dataset-simple.json", true);
+      wrap.deleteDataset(datasetSummaryModel.getId());
+      datasetSuccess = true;
+    } catch (DataRepoUnauthorizedClientException ex) {
+      datasetSuccess = false;
+    }
+    logger.info("Role {} dataset {}", role, datasetSuccess);
+
+    // Test for ability to add a user to the user role of the profile
+    boolean addUserSuccess;
+    try {
+      wrap.addProfilePolicyMember(profileId, "user", userUser.userEmail);
+      addUserSuccess = true;
+    } catch (DataRepoNotFoundClientException | DataRepoUnauthorizedClientException ex) {
+      addUserSuccess = false;
+    }
+    logger.info("Role {} addUser {}", role, addUserSuccess);
+
+    // Test for ability to delete a user from the user role of the profile
+    boolean deleteUserSuccess = addUserSuccess;
+    if (addUserSuccess) {
+      try {
+        wrap.deleteProfilePolicyMember(profileId, "user", userUser.userEmail);
+        deleteUserSuccess = true;
+      } catch (DataRepoNotFoundClientException | DataRepoUnauthorizedClientException ex) {
+        deleteUserSuccess = false;
+      }
+    }
+    logger.info("Role {} deleteUser {}", role, deleteUserSuccess);
+
+    // Test for ability to retrieved the policies of a profile
+    boolean retrievePolicySuccess;
+    try {
+      wrap.retrieveProfilePolicies(profileId);
+      retrievePolicySuccess = true;
+    } catch (DataRepoNotFoundClientException | DataRepoUnauthorizedClientException ex) {
+      retrievePolicySuccess = false;
+    }
+    logger.info("Role {} retievePolicies {}", role, retrievePolicySuccess);
+
+    switch (role) {
+      case NONE:
+        // all operations should fail
+        assertThat(
+            "NONE all operations should fail",
+            !(foundEnumeration
+                || retrieveSuccess
+                || datasetSuccess
+                || addUserSuccess
+                || deleteUserSuccess
+                || retrievePolicySuccess));
+        break;
+
+      case USER:
+        assertThat(
+            "USER operations that should succeed",
+            (foundEnumeration && retrieveSuccess && datasetSuccess));
+        assertThat(
+            "USER operations that should fail",
+            !(addUserSuccess || deleteUserSuccess || retrievePolicySuccess));
+        break;
+
+      case OWNER:
+        assertThat(
+            "OWNER all operations should succeed",
+            (foundEnumeration
+                && retrieveSuccess
+                && datasetSuccess
+                && addUserSuccess
+                && deleteUserSuccess
+                && retrievePolicySuccess));
+        break;
+    }
   }
 }
